// client/src/utils/constants.ts
export enum AuthDataKeys {
    USER_ID = 'userId',
    ACCESS_TOKEN = 'accessToken',
}
export enum AuthResult {
    NONE = 'none',
    SUCCESS = 'success',
    ERROR = 'error',
}

export enum FormNames {
    SIGN_UP = 'sign-up-form',
    SIGN_IN = 'sign-in-form',
    ADD_WORKSPACE = 'add-workspace',
}
export enum UserStoreDataKeys {
    ID = 'id',
    EMAIL = 'email',
    PASSWORD = 'password',
}

export enum ResponseErrorMessages {
    UNAUTHORIZED_MESSAGE = '',
}